requirements.txt
openai==0.27.8
python-dotenv==1.0.0
questionary==1.10.0
rich==13.4.1
